+++ Pandora/Paradigm/Controlflow/Effect.hs
@@ -1,7 +1,6 @@
 module Pandora.Paradigm.Controlflow.Effect (module Exports) where
 
 import Pandora.Paradigm.Schemes as Exports
-import Pandora.Paradigm.Controlflow.Effect.Schematic as Exports
 import Pandora.Paradigm.Controlflow.Effect.Transformer as Exports
 import Pandora.Paradigm.Controlflow.Effect.Interpreted as Exports
 import Pandora.Paradigm.Controlflow.Effect.Adaptable as Exports