+++ Pandora/Paradigm/Controlflow/Effect/Interpreted.hs
@@ -1,4 +1,6 @@
-module Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (..)) where
+module Pandora.Paradigm.Controlflow.Effect.Interpreted (Schematic, Interpreted (..)) where
+
+type family Schematic (c :: (* -> *) -> k) (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u
 
 class Interpreted t where
 	{-# MINIMAL run #-}