+++ Pandora/Paradigm/Controlflow/Effect/Schematic.hs
@@ -1,3 +1,0 @@
-module Pandora.Paradigm.Controlflow.Effect.Schematic (Schematic) where
-
-type family Schematic (c :: (* -> *) -> k) (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u