+++ CHANGELOG.md
@@ -264,3 +264,4 @@
 * Define `Set` interface for data structures with `member` method
 * Rewrite `zoom` function with right adjunction and move it to `Inventory` module
 * Move `Adjoint` instance for `(:*:)` and `(->)` to `Paradigm.Primary.Functor` module
+* Remove `Schematic` module and move `Schematic` type family to `Interpreted` module